Refactor PDF and playground to share code
