+++ src/docgen/pdf/react-pdf/pdf-page/pdf-page.tsx
@@ -1,64 +1,19 @@
 import React from 'react';
 import { Page, View, StyleSheet } from '@react-pdf/renderer';
-import Html from 'react-pdf-html';
+import { PdfHtmlBlock } from "./pdf-html-block";
 import { PdfFooter } from './pdf-footer/pdf-footer';
-import * as styles from '../../../../styles/variables/js/style-variables';
+import { pdfStyleSheet } from '../pdf-styles/pdf-styles';
 
-const fontSize = 10;
-
-const pdfStyleSheet = StyleSheet.create({
-  page: {
-    padding: '1cm',
-    fontSize,
-    margin: 0,
-  },
-});
-
-const styleInfo = {
-  color: styles.ColorTextInfo,
-  backgroundColor: styles.ColorBackgroundInfo,
-  borderLeft: `5px solid ${styles.ColorBorderInfo}`,
-  padding: styles.SizeMessagePadding,
-  //borderRadius: styles.SizeMessageBorderRadius,
-}
-
-const em = (em: number, relativeSize: number = fontSize) => em * relativeSize;
-
-const htmlStyleSheet = {
-  p: {
-    color: styles.ColorText
-  },
-  h1: {
-    fontSize: em(1.4),
-    marginVertical: em(0.2, em(2)),
-    fontWeight: 'bold',
-  },
-  blockquote: styleInfo,
-  ['.w-information']: styleInfo,
-  pre: {
-    //padding: 12px,
-    //background-color: #f3f3f3,
-    borderLeft: `1px solid ${styles.ColorBorder}`,
-    //line-height: 14px;
-    //border-radius: 6px;
-  }
-};
+const reactPdfStyles = StyleSheet.create(pdfStyleSheet);
 
 export const PdfPage = ({
   page,
   parameters
 }) => {
   return (
-    <Page size="A4" style={pdfStyleSheet.page}>
+    <Page size="A4" style={reactPdfStyles.page}>
       <View>
-        <
-          Html
-          resetStyles
-          style={{ fontSize }}
-          stylesheet={htmlStyleSheet}
-        >
-          {page}
-        </Html>
+        <PdfHtmlBlock page={page} />
       </View>
       <PdfFooter parameters={parameters}/>s
     </Page>
